[FIX] account_reports: use sudo() to compute journal's display name
This commit is the counterpart of a commit in community to add a test.

Steps to reproduce:
- Create a branch for a company
- Create a journal on the parent company with a different currency from the company
- Create a user that can access both the parent company and the branch, but can't change the settings of Odoo
- Log as this user, select only the branch as current company
- Accounting > Reporting > Profit & Loss
- Access Error

Cause:
`_compute_display_name` on the journal reads `journal.company_id.currency_id` without sudo.
The current user cannot read on `company_id` because of the rule `res_company_rule_employee`.

Solution:
Use `sudo()` to read the currency of the company.

opw-4847500

closes odoo/enterprise#89018

X-original-commit: 5a5b8876e23413badbd0e6450e1161b812ad7f7c
Related: odoo/odoo#216673
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>