[FIX] sale_subscription: prevent traceback when next_invoice_date_at_resume is false
When a quotation is created and the user attempts to pay via the portal for a
subscription where plan_id.pausable_by_user is False, the
next_invoice_date_at_resume field can be None. This led to a crash when the
template attempted to call format_date(None), resulting in a 'NoneType' object
is not callable error.
This commit adds a safe check in both the controller and the template to ensure
the date exists before attempting to format it.

closes odoo/enterprise#88571

X-original-commit: b57073701895c785d7bf839137b5bce661969f64
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gurupreet Singh (gusi) <gusi@odoo.com>