[FIX] account_bank_statement_import_csv: importing only the first batch
**PROBLEM**
When importing bank statements from the dashboard, it only imports the first batch (by default the first 2000 lines) instead of importing the whole file.
This is inconsistent with the behavior of the import done from the reconcilation page.
Also, on the import page, the file name is incorrect (it's always `bank_statement_import.csv`)

**STEP TO REPRODUCE**
1. install the accounting module
2. goes on the dashboard, click on the 3-dot button on the kanban for the bank account, and import a file.
3. make sure the file will be imported in multiple batches (reduced the batch size if necessary) and click on import.
4. notice how only the first batch was imported.

**CAUSE**
In python,
The `AccountBankStmtImportCSV` class override the execute_import method of the `base_import.import`.
In this override, we add a entry in the `messages` list.
(see `enterprise/account_bank_statement_import_csv/models/account_bank_statement_import_csv.py`)

In JS,
all entry in messages are treated as errors, and the import is interrupted.
(see `odoo/addons/base_import/static/src/import_model.js`)

The message entry added in the python was used in the past to automatically open the reconcillation page with the statement lines added.
This feature was removed, but not the message.

**FIX**
Remove the problematic message entry.

opw-4823808

closes odoo/enterprise#91305

X-original-commit: d621dbf345218ed3f114b5b6294f5912219f326f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Léo Gizard (legi) <legi@odoo.com>