[FIX] stock_barcode: scan packaging to update already existing lines
Issue:

When processing a picking form the barcode, uncompleted lines refering
to a product are not updated by scanning a different packaging.

Steps to reproduce:
- In the settings enable "Units of Measure & Packagings"
- Create a product using the base Unit as uom and
  allowing pack of 6 aspackaging (optional)
- Inventory > Configuration > Product > Units & Pacakagings
- Click on pack of 6 > Packaging Barcodes > New
- Create a new one for your product
- Create and confirma  receipt for 6 units of your product
- Go to the barcode app and scan a pack of 6

> A new line 1 pack of 6 was created rather than updating the already
>  existing one.

Cause of the issue:

When the packaging is scanned, we will call the `findLine` of the
`BarcodeModel` in order to determine if there is already an existing
line that could absorb the scanned qty. However, this call currently
requires the uom of the barcodeData to coincide either with the line uom
or its referred packaging to be a valid candidate:
https://github.com/odoo/enterprise/blob/fb405cee9b183d697e3cf0e33bb49371b87cb1ac/stock_barcode/static/src/models/barcode_model.js#L1748-L1753

Note:

This additional check was added in a9a20f8ec5a7883a08eb4bea15dbf3c4cc97caac
when the uom and packagings were refactored and merged together.
However, it is really unecessary since the system can handle different
uoms via these lines:
https://github.com/odoo/enterprise/blob/fb405cee9b183d697e3cf0e33bb49371b87cb1ac/stock_barcode/static/src/models/barcode_model.js#L1450-L1466

opw-4866981

closes odoo/enterprise#90059

X-original-commit: d5b34949e983714c7bf425ea7b31b9d83191f812
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>