[FIX] voip: resolve non-blocking errors on click
UI-wise, there are two types of VoIP errors: blocking and non-blocking.
Blocking errors are unrecoverable, they're displayed at the top of the
UI and prevent further use of the softphone. Non-blocking errors, on the
other hand, can be dismissed by a simple click on the error screen.

Problem: the handler that hides the error screen in case of non-blocking
errors wasn't attached to anything, and even if it was, its
implementation was wrong! This commit makes it work as intended.

Part of task-4891947

closes odoo/enterprise#88437

X-original-commit: fa284901468911108210c9d4a5a6934188bbcc6d
Signed-off-by: Mohammed Basioni (basm) <basm@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>