[FIX] pos_appointment: compute capacity booked
Currently when making an appointment for x amount of people, inside the
restaurant it will show 0 as the number of people who booked.

Steps to reproduce:
-------------------
* Activate booking on the restaurant
* Open the booking preview
* Place a reservation for 5 people
* In the restaurant and remove default filter
> Observation: It shows 0p on the reservation

Why the fix:
------------
By having a default value on the field is would not trigger the compute unless
a change was made on `resource_ids` after the create was over.

After fixing this bug we also aboserve that if we make a reservation for 1
person but all the tables in the restaurant are 4 person seatings, the
reservation will show 4p.

The reservation should in priority depend on the number entered while booking
otherwise we can compute it regarding the resources.

opw-4882681

closes odoo/enterprise#90421

X-original-commit: 94b5941b56d4cb36bb6fd290551fa002fa8e37d3
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>