[IMP] website_sale_renting : improve datetime picker frontend design
This commit fine-tunes the design of the datetime picker within the
front-end after the merge of Commit[1].

Prior Commit[1], the previous, next, and month selection buttons were all
placed in a flex at the top of the datetime picker. After Commit[1],
buttons are re-ordered in a different layout, with the right button class
on each of them.

After this change, each button is separated and thus has its radius and
background. In the backend, this `btn-light` is okay since it matches the
background of our views, while in the front-end, this color is dynamic
and does not match the page background. As each button is separated, it
results in a weird visual result.

To fix this issue, we remove these `btn-light` classes and add a `btn-sm`
one to handle the size of these buttons.

This change is meant to affect the datetime picker occurrences within the
front-end but does not affect the `web_editor` one, as it inherits the
backend design.

task-4893742

[1]: e2b78fca435f268515f9593c2af821616aff7ebb

closes odoo/enterprise#88480

Signed-off-by: Valentin Chevalier <vcr@odoo.com>