[FIX] ai: fix AI-draft chat window bugs
Before this commit, ai actions like "insert", "send message" and "log
note" were appearing on both the messages sent by the AI and the user's
own messages. Of course, the actions shouldn't appear in the user's
messages.

To fix this, we changed one element in the conditions for these actions.
Prior we were fetching the user_id from the author of the message and
comparing it with the current user. Recent changes make it impossible to
fetch the user_id like so. Now we use the function isSelfAuthored which
would return false for AI messages and true for the user's own messages.

Also another bug fixed in the process, the copy action button was placed
under message bubbles when inside the discuss app.

Additionally, we extend the thread model method `computeCorrespondent`.
We do this in order to remove the out of office alert that is shown if
the current user is on time off and they have a conversation with AI. In
AI chats, the correspondent was the user itself which didn't make sense
and so now the value should be empty.

Finally we added tests, that should keep the AI flow stable in case of
future changes.

Task-4895064

closes odoo/enterprise#90949

X-original-commit: 66bd626e11840efa25c097f0e5965b288c9a73f6
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>