[FIX] appointment: allow user to create appointment through calendar event
When booking an appointment that uses resources, users in the
Appointment/User group may encounter the following access error:

    You are not allowed to modify 'Appointment Booking Line' (appointment.booking.line) records.
    This operation is allowed for the following groups:
        - Appointment/Administrator

This occurs when exactly one resource is set on the appointment type.

In that case, the invisible and read-only `appointment_resource_id` field on
the `calendar.event` form is not sent by the browser during web_save.
Odoo then computes its value server-side. Because this field shares its
inverse method with `resource_total_capacity_reserved`, and that field is also
involved in the save, the inverse is triggered.

Since `appointment_resource_id` is now truthy, the inverse method tries to
write to `booking_line_ids` using the computed resource. However, users in
the Appointment/User group lack write access on `appointment.booking.line`,
leading to an access error.

To fix this, we now sudo the logic in the inverse method, as already done in
`_inverse_resource_ids_or_capacity`. This avoids the access error while
preserving expected behavior.

opw-4614976

closes odoo/enterprise#92186

X-original-commit: e37284d3b99e4f8671c83922d72f12b2d9ae9e33
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>