[IMP] documents{_spreadsheet}: add action buttons in list view
This task adds several action buttons on each row of the
list view to ease access to those actions.

Those action buttons are visible when hovering over row.

Known limitation, the widget's column size is not dynamically
evaluated based on its content, so we must define a fixed size
suitable for the max number of actions that can be shown.

Task-4794981

Part-of: odoo/enterprise#88509
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Co-authored-by: Florian Charlier <flch@odoo.com>