[FIX] account_reports: vies vat check with missing base_vat dependency
the field vat_check_vies on res.company is added in the base_vat module, which
is not a required dependency for account_reports. the function
_generic_vies_vat_check might therefore result in an error with a missing field.

To reproduce the error, create a company for a european country that doesn't
depends on base_vat, for example Estonia (l10n_ee), and open the returns page.

runbot issue: 227603

closes odoo/enterprise#88741

X-original-commit: db582b88d10cc5999622041961a3dbd5a5b72e0d
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>