[FIX] sign: Sign documents download fixes
This commit reverts to the previous behavior where
the certificate of completion is never merged with
the original PDF, but is instead downloaded as a
separate file.
It also handles the downloading of completed/
original documents in the list form/view:

- For users who are involved in a sign request
(either as a signer or CC'd), the original documents
are now available for download. They are now granted
read access to the related sign documents.
- This ensures that any user who is asked to sign
or is CC'd on a sign request can view and download
the original documents, without exposing
templates or documents to unrelated users.

closes odoo/enterprise#90551

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Abdelrahman Shaaban (ABDH) <abdh@odoo.com>