[FIX] project_account_budget: adapt project analytics to budget lines
In version 17.3, the analytic mixin was introduced in budget.line model to allow to link a budget line to multiple analytic accounts belonging to different analytic plans.
The issue is that we have not adapted the code in project profitability, to take those changes into account. A problematic situation could be the following:

1. We have a project where the account "X" is set to the main project plan (the first one) "Projects" BUT the account "X" actually belongs to the analytic plan "Departments"
(yes, it's possible that there is a mismatch between the project plan and the analytic plan, especially after the 18.0 upgrade where "analytic_account_id" just became "account_id", no matter in which plan it was).

2. We have a budget line linked to the account "X" for the plan "Departments".

3. We want to display the budget lines in project profitability, but this line does not appear. That's because we currently only look for the budget lines belonging to the analytic plan "Projects" aka the main/first plan, that match the account "X" of the project.
But as account "X" actually belongs to the analytic plan "Departments", no budget lines are found.

task-4901400
version-18.0

closes odoo/enterprise#90554

X-original-commit: b375390bb5e7f5fd8e5c71ecfa22d8ddbfe574b0
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Maxime de Neuville (mane) <mane@odoo.com>