[IMP] *: update the res.users form views
This commit modifies the fields on the preferences / profile form, which
allows users to read or modify their own data depending on their
permissions, and on the user form for the administrators. The aim of this
change is to display more relevant information to users.

As some fields are removed from the forms and are not longer used, they are
also removed from the res.users model.

Community PR: https://github.com/odoo/odoo/pull/215789
Upgrade PR: https://github.com/odoo/upgrade/pull/8207

Task-4775990
Revert "For form user"

This reverts commit 11b14cc20379a1bf70724577875e7ba2096cbe34.

For form

Part-of: odoo/enterprise#88518
Related: odoo/odoo#215789
Related: odoo/upgrade#8207
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>