[FIX] iot: prevent printer selection being cancelled
Steps to reproduce:
1. Connect IoT Box and any printer that accepts PDF
2. Turn on Reception Report option on Inventory Settings
3. Set configuration of Receipts to print out Reception Report and Label
4. Assign Reception Report and label to the printer
5. Create a PO and run through the Reception process
   (PO > Reception of Delivery)
6. Validate the Reception of the order

-> Result:
   Odoo will prompt the customer to select a printer, however we
   are not able to choose a printer in time as the process continues
   without selecting one. This in return does not send the report to the
   printer via IoT.

The root cause of this bug is that the IoT report handler JS function
returns too early, it resolves once the printer selection popup has
appeared, instead of resolving once the printer has actually been
selected and is starting to print. Therefore the multi-report printing
code assumes the print is done and triggers the next print, which causes
the popup to close before the user can select a printer.

To solve this bug, we listen for a 'printer-selected' event in the
handler, and resolve only once we have received this event. This fixes
the flow, allowing a printer to be selected for each report that is
being printed in sequence.

opw-4790299

closes odoo/enterprise#89756

X-original-commit: 697c481825b09e5eba4d4714c7a8d87825a575eb
Related: odoo/odoo#217944
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>