[IMP] knowledge: redirect non-internal users to public views
In practice, when an internal user edits an article and wants to share it,
they often share the URL from their browser's address bar rather than using
the share links in the share panel.

The issue is that external users who click this link cannot access the article
via the backend. They are hence unable to access the article even though they
could access it via the Knowledge portal or public view.

To address this, a redirection to the main backend controller has been implemented
that automatically sends non-internal users attempting to access the Knowledge
backend view to the Knowledge portal or public view. With this redirection, the
backend link also becomes functional for them.

task-4797702

closes odoo/enterprise#88541

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>