[FIX] documents: restore document access from url
Reproduce:
1. Select a document so that its token is now part of the URL
2. Refresh the page
3. The document is not selected as it should

This is problematic because if you are given a link to access
a document in a folder with any siblings, you have no way to
know which one the link pointed to.

Technically, since 18.3, records are no longer found in
`onMounted`.
Note that due to another (older, independent, and less
critical) bug, non-previewable records are not focused.

Task-4873715

closes odoo/enterprise#88875

X-original-commit: f37305a518e481a101c6b0bf4c440426aea9b339
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>