[FIX] test_l10n_be_hr_payroll_account: fix wrong time off type
This commit fixes an issue in the demo data. A time off allocation was done using a time off type that is not in Belgium. This caused access error when trying to access the employee.
This commit fixes the issue by changing the time off type to a belgian one.

task-4896664

closes odoo/enterprise#89008

X-original-commit: 11e092783f0b03f086ecd899b4f88abfcaf5faab
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>