[IMP] sign: set model's partner_id as default signer
Before this commit, when using the "Signature Request" action from a form view,
The current user was set as the default signer (customer role).

In this commit, if the record has a related partner_id, it will now be used as
the default signer instead.

This improves the behavior introduced by automatically selecting the correct
partner from the record.

task-4246455

Part-of: odoo/enterprise#88774
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>