[IMP] iot: add webRTC support
This PR adds a third connection type to the IoT box, using the
WebRTC protocol. This commit handles the browser side of the connection.

The WebRTC connection is added as a third connection type to the
existing `iot_http` service, with the highest priority. If it fails for
any reason it can still fallback to longpolling then websocket.

We use the Browser APIs `RTCPeerConnection` and `RTCDataChannel` to
setup the WebRTC connection to the IoT box. The flow is as follows:

1. Browser creates an `RTCPeerConnection` instance, generates an 'offer'
   and sends it to the IoT box via the existing WebSocket connection.
2. The IoT box receives this offer, and sends us back an 'answer', again
   via the WebSocket connection.
3. At this point the browser will automatically establish the connection
   to the IoT box. Any future calls to `iot_http` will use this
   connection (unless it fails, in which case it will fallback to the
   other connection types as before).

This implementation is intentionally very simple.
This means we do not use a STUN server, and so the connection will only
work on the local network. This isn't a huge problem because in the case
you are not on the same network as the IoT box, the WebSocket connection
will still work and there would be little gain from using WebRTC.

task-4894918

closes odoo/enterprise#90179

Related: odoo/odoo#218871
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>