[REF] documents: apply formatter in documents_draggable and hooks
Only formatting, no other changes here.

The module is listed for the fomratting pre-commit hook but it looks
like recent commits were made without it.

In order to make the next commit more readable and avoid the trigger of
the formatter there, it is done in this commit.

Part-of: odoo/enterprise#88620
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>