[REF] web_enterprise, *: use textContent instead of innerHTML (empty)
Writing on innerHTML is generally not recommended as it can lead to
security issues. This is highlighted by triggering the ci/security.

`setElementContent` utils is now available to safely write it on
`innerHTML` by checking the content type (markup in particular).

However, when writing an empty value, using `textContent` has the same
effect of clearing all children. It is therefore preferable to use it
instead, in order to remove direct usage of `innerHTML` in code and to
avoid triggering unncessary security checks.

https://developer.mozilla.org/en-US/docs/Web/API/Node/textContent

closes odoo/enterprise#88628

Related: odoo/odoo#216066
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>