[FIX] hr_referral, account_loans: fix translation issue
Issue:
Prior to this commit, a translation issue occurred due to the use of a list comprehension.
The _get_translation_source function attempts to scan the local variables,
but in the context of a list comprehension, only variables defined
within the comprehension are accessible. As a result,
variables like uuid and cursor were not available to the _get_lang function,
ultimately leading to an error.

Fix:
Replaced the list comprehension with a standard for loop to ensure
proper access to local variables.

runbot-98198

closes odoo/enterprise#89719

X-original-commit: 3bff53dc1fc57f571917e8728c6d8f8110ef4620
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>