[IMP] hr_payroll: pay run, payslip generation
Problem
----------
Generation of payslip with valid_versions is not enough flexible
With work entries, we have to generate payslips for employees.

Objective
----------
Change the _get_valid_versions_domain to _get_valid_versions to get only ids
Adapt the signature of generate_payslips to use employees or versions to generate payslips

Solution
----------
Now the _get_valid_versions give the valid versions ids, so we can directly browse them or add them in a domain, the domain is not forced
The generate_payslips can work with versions (first argument by default) and employees, if employee_ids are given, this will fetch valid versions for these employee depending on the pay run config directly.
So via the front, we can call generate_payslips with a list of employees if we don't have versions directly. (it will automatically select all contracts valid in the range of payrun as before with contracts)

task-4900019

closes odoo/enterprise#89174

X-original-commit: 5c553228e8cb2320eadc0e530365a00661bea43e
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Mathias Lebel (lebm) <lebm@odoo.com>