[REF] web_tour: uncomment beforeUnload listener
In this commit, we uncomment the listener on beforeUnload to ensure that
expectUnloadPage is correctly used in tours.
From now on, if a step produces a page unload but didn't declare it,
the tour will fail with an Error. This forces step producing page
reload/redirect to declare them, s.t. the tour stops after that step,
waiting for the next page to be loaded to continue.
This is done to remove non determinism in tours.

closes odoo/enterprise#88717

Related: odoo/odoo#216213
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>