[IMP] {website_}appointment: enable template selection at creation
In order to give the user the possibility to reuse the
appointment type templates that we offer to choose from
when the action helper is displayed (when no record is
found), we enable the template selection when creating
a new type.

We add a new AppointmentTemplatePickerDialog component
that will be opened when using the 'New' button in the
form, list and kanban views.

We use the same template card in both the action helper
and the template picker at creation. (It is therefore
parametrized and a few classes are customized in the
templates of both components when calling the card
template). The styling is adapted accordingly.

A tour is adapted accordingly.

Task-4898893

Part-of: odoo/enterprise#88637
Related: odoo/upgrade#8014
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>