[FIX] account_avatax: fix _check_address translation issue
Issue:

Prior to this commit, a translation issue occurred due to the use of a
list comprehension.  The `_get_translation_source` function attempts
to scan the local variables, but in the context of a list
comprehension, only variables defined within the comprehension are
accessible. As such, the variables `_get_lang` ultimately looks for to
try and retrieve the current language are not in the stackframe we're
looking at, ultimately leading to a warning.

Fix:

Use `Environment._` as it doesn't do stackframe inspection and thus
does not suffer from this issue.

runbot-159883

closes odoo/enterprise#90216

X-original-commit: 398c82a7dbaa629e6bbee85b741818df81f1115d
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>