[FIX] account_online_synchronization: Prevent blocking error on currency mismatch
This commit avoids blocking users when connecting a journal to an online account
that has a predefined currency.

Previously, if the journal or its default bank account had entries in a different
currency than the one sent by Odoo Fin, the process would fail with an error,
preventing completion.

Now, the system silently ignores the currency mismatch error and skips setting the
currency, allowing the synchronization to proceed.

opw-4751616

closes odoo/enterprise#89214

X-original-commit: be4bee5c2491bb5bb18fed754cb246c8083fac6c
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>