[IMP] account_accountant: account type
In this commit: https://github.com/odoo/enterprise/commit/c0f8c5d17b9d78b640d744d8bfb6a2bb4002776a
We changed a condition to have the possibility to match move with an account
that allows the reconciliation but by doing so, we authorize moves with a cash
line to be match. This will trigger a constraint since we cannot have multiple
cash or bank accounts on the same statement move.
To avoid that we add a condition that the type of the account can be reconcile
but is not an asset_cash or liability_credit_card

no task id

closes odoo/enterprise#89222

X-original-commit: 6a3cb2b9a85f9cec90aaeacf0cc37ba376861c32
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>