[FIX] documents_account: sync xml document on attachment creation
Event with the document integration correctly setup, e-invoice xml
generated in the send&print wizard will not create an associated
document unless the attachment is actually sent via mail

Steps to reproduce:
- Have a EU Company setup
- Enable and configure Peppol Electronic Invoicing
- Enable documents integration with accounting
- Create an invoice to a Peppol enabled customer, confirm
- Open send&print wizard, enable only 'Download' and 'BIS Billing 3.0'

Issue: Document related to the xml attachment is not created. This will
work as expected when the message is sent to the customer

It occurs because, in case of xml attachment, we don't create an
associated document until some attachment have been registered on the
move. However, the one2many field on the move (`attachment_ids`) does
not immediately register the new attachment as soon as it
is created, effectively delaying document creation until the mail attachments
are created to be sent

opw-4720588

closes odoo/enterprise#90973

X-original-commit: db8cb7ede8a40509d4f439c1ea69523214580b0c
Related: odoo/odoo#220471
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>