[FIX] account_reports_cash_basis: nightly build failing
The builds were failing due to this PR: https://github.com/odoo/enterprise/pull/72675

The reason is that we removed one column of the report.
Another issue is that a test was trying to call an old expand function which is not in use anymore.

Build failing: https://runbot.odoo.com/odoo/runbot.build.error/226793/runbot.build.error/226793

closes odoo/enterprise#88885

X-original-commit: 3fc56795fc9c9104855ecc9985cd0370607321d1
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>