[FIX] sale_commission: prevent error when clearing date_to in commission plan
Currently, an error occurs when clear the date_to field from an already created
record.

Steps to Reproduce:
 - Install the sale_commission module.
 - Go to Commission Plans and create a record.
 - Now, clear the date_from field and save.

TypeError: '>' not supported between instances of 'bool' and 'datetime.date'

This error occurs when we remove the date_to field from an already created
record. The error started occurring after
https://github.com/odoo/enterprise/commit/926c8f041b51a04f0153be27cfcc8658a07be5b0
this commit that introduced the daterange widget for the date_from field.
And when we clear the date_to field and try to save, it triggers the constraint
[1] and raises the error.

[1] https://github.com/odoo/enterprise/blob/377321fd7e7b1b2026734b75c79c74c8947e2e6c/sale_commission/model/commission_plan.py#L64

This commit ensures that the record is saved only if the date_to field is also
present in the record.

sentry-6710699656

closes odoo/enterprise#91059

X-original-commit: 099ce35a6e3e8eaddc4c1444fb50ea4160dcb6c5
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Ashutosh Sharma (assh) <assh@odoo.com>