[IMP] appointment,*: change manual confirmation to auto_confirm
*appointment_account_payment,pos_restaurant_appointment

Inverse the definition of the field appointment_manual_confirmation
into a new auto_confirm field, with default = True. Labels are
changed accordingly.

Before, we could activate a manual confirmation, setting the
booking status to 'request' when booking over a certain percentage
of the total capacity.

Now, we have by default a true auto_confirm, and when true we show
the treshold option. The percentage value is now 100% by default,
and we still only show it when manage_capacity is True.

This means that when manage_capacity is True, the only way to always
auto confirm the meetings is to have 'auto-confirm' set to True, and
a 100% manaul_confirmation_percentage. A small boolean computed field
is added to know whether we always confirm or not depending on the
configuration.

UPG PR: odoo/upgrade#8014

Task-4898893

Part-of: odoo/enterprise#88637
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>