[FIX] l10n_se_sie4_import: Ignore object list for transactions
The aim of this commit is ignoring the whole object list when
we import transactions. The object list is represented by
the elements between {} [[1]]. As we don't need these elements,
we do ignore them to avoid a potential issue where we identify
an element from this object list as the transaction's balance.

opw-4868415

[1]: https://sie.se/wp-content/uploads/2020/05/SIE_filformat_ver_4B_ENGLISH.pdf 8.21

closes odoo/enterprise#89252

X-original-commit: 5a29f5aac0f677b7d3497ff991b0847094967a14
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>