[IMP] appointment,*: split assign_method in two fields
*appointment{_account_payment,_crm,_hr{_recruitment}},
appointment_google_reserve, pos{_restaurant}_appointment,
test_appointment_full, website_appointment

In order to make the option clearer, we split assign_method
parameters: the ordering of selection (datetime or user /
resource first) and the manual or auto assignment. As each
only has two options, we introduce two stored boolean fields.
In order to be explicit on the options, we use computed
selection editable fields in order to use the radio widget
with labels. This could be removed if in the future we have
access to a radio boolean widget with translatable labels.

The default values match the previous resource_time value.

Of course, as we only worry about selection order when manual
selection is enabled, we hide the field on the form when
the assignment_method is 'auto'. Note that the new is_auto_assign
value should always be checked first when checking the new
is_date_first value. This way, we do not need it to depend
on is_auto_assign as its value will simply be ignored.

Update various occurences accordingly

UPG PR: odoo/upgrade#8014

Task-4898893

Part-of: odoo/enterprise#88637
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>