[REV] website_sale_renting: revert unwanted props within the datepicker
This reverts commit[1].

The goal of Commit[1] was to introduce more flexibility to the datetime
picker by using props to handle the design of the component within the
frontend, to preserve the design introduced with the revamp of the
component within the backend while fine-tuning its design within the
frontend.

During the review process, it was decided that changes would be applied
on the main component directly instead of monkey patching the component,
which invalidated the need of props.

As the counterpart was not merged, this change shouldn't have landed
as well.

[1]: e650409bd8b027074f262dfdfc407d8d71bb8e36

closes odoo/enterprise#88806

Signed-off-by: Valentin Chevalier <vcr@odoo.com>