[FIX] sale_commission: don't duplicate lines when no achievement
Before this commit, when the user grouped by quarter, some periods would
not be merged.

Example:
January: no achievement
February: one achievement
March: no achievement

The first quarter would have two lines: one with achieved 0 (January, March) and one with
the amount of february.

It happend because we would group by achievement currency. It is NULL
for January and March and therefore it can't be merged with the rest.

task-4783929

X-original-commit: 270c18be8bf5ee89ec17d313013e0f168ad375e0
Part-of: odoo/enterprise#91183
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>