[FIX] account_invoice_extract: don't duplicate partners
When the OCR parses documents from the same supplier concurrently, it
can happen that it creates duplicated `res.partner` entry for it.
This typically happen when the update of the documents is done through
the webhook.
In that case, the different SQL transactions overlap each other and
can't "see" the partner created by another request.

In order to prevent this, a new field `is_created_by_ocr` and a unique
index are introduced to be able to detect duplicated partners created by
the OCR.

This solution is quite costly as it requires introducing a new field and
an index on `res.partner`...
Other alternatives were considered:
- Updating the documents through a cron instead -> introduces latency,
  especially on the SaaS where we can't force a cron to run immediately,
  even with a cron trigger.
- Introducing a new model on which we could add the uniqueness model ->
  also costly, requires a new model just to fix the issue, and we need a
  way to frequently clean the records, so probably a cron on top of it
  (or using autovacuum, but its frequency might be too low).
- Adding random delay on OCR side before calling the webhook ->
  introduces latency and no guarantee, it only reduces the chances of
  duplicated partners.
- Using explicit SQL locks -> no guarantee, we only know what should be
  locked late in the SQL transaction (after fetching the results from
  the OCR), it wouldn't be very effective. Locking the whole
  `res_partner` table from the start of the transaction would be too
  restrictive and equivalent to serializing the transactions.

task-4904333

closes odoo/enterprise#89410

Helped-by: Julien Castiaux <juc@odoo.com>
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>