[FIX] industry_fsm_stock: disable locking of SOs for FSM tasks
Issue:
When the "Lock Confirmed Sales" setting is enabled, adding a product to
a Field Service task would lock the associated sales order before its
picking could be created. The confirmation process locks the order first
, and the subsequent step to create stock moves
(`_action_launch_stock_rule`) explicitly skips locked orders.

Steps to reproduce:
1. In Sales > Configuration > Settings, enable "Lock Confirmed Sales".
2. Go to a Field Service task.
3. In the kanban view via the "Products" smart button, add a product.
4. Come back to the task, refresh the browser, open the linked Sales
Order from the smart button.
5. **Before this fix:** The SO is locked, no picking is created, and SO
lines will have delivered_qty set even though we haven't delivered
anything.
6. **After this fix:** The SO is not locked, and a Delivery smart button
is correctly displayed, linking to the generated picking.

This behavior left the sales order in an inconsistent state: it was
locked, but had no corresponding delivery order. As a result, users were
blocked from removing or modifying the products added to the task, as
the system prevented edits on a locked order with no picking to cancel.

Let's step through what's going on here;

0. We enable "Lock Confirmed Sales" option.
1. Add a product to the Field Service(FS) task
https://github.com/odoo/enterprise/blob/4256ea170c31209b83ff808e51145757571f8556/industry_fsm_sale/controllers/catalog.py#L19-L42
2. set_fsm_quantity() is invoked, which will create a SO for the FS task
https://github.com/odoo/enterprise/blob/35e1f47160d5840e8701fc0ad43ec42bbfaf9e36/industry_fsm_sale/models/product_product.py#L144
3. `_fsm_create_sale_order()` will immediately `action_confirm()` the new SO, because of the reasons in the function description.
https://github.com/odoo/enterprise/blob/2347dac2568dbcf83578f044f8fe41ab70a20f47/industry_fsm_stock/models/project_task.py#L124-L134
https://github.com/odoo/enterprise/blob/2347dac2568dbcf83578f044f8fe41ab70a20f47/industry_fsm_stock/models/project_task.py#L134
4. `odoo/addons/sale/models/sale_order.py/action_confirm()` will lock SO because we enabled "Lock Confirmed Sales"
https://github.com/odoo/odoo/blob/863c064fd911cb4eeedad0abec82adb4690128f6/addons/sale/models/sale_order.py#L964
... other operations ...
5. `_action_launch_stock_rule()` is invoked, is in charge of creating Pickings for SOs
https://github.com/odoo/odoo/blob/3be83d363786526fe470e41122169e594b5467fb/addons/sale_stock/models/sale_order_line.py#L303
https://github.com/odoo/odoo/blob/3be83d363786526fe470e41122169e594b5467fb/addons/sale_stock/models/sale_order_line.py#L343-L344
6. But, since `line.order_id.locked` , the system do not create Pickings for the SOs.
https://github.com/odoo/odoo/blob/3be83d363786526fe470e41122169e594b5467fb/addons/sale_stock/models/sale_order_line.py#L315-L316
The line was introduced by
https://github.com/odoo/odoo/commit/17bece3e797913bcba8dd7e07fc8541c0a45e3f7

This commit resolves the issue by passing a context key
`fsm_create_sale_order=True` when confirming a sales order that is being
created from a Field Service task. The `_should_be_locked` method on the
sale order is overridden to check for this context key. If the key is
present, it prevents the order from being locked within that specific
transaction, allowing the stock rules to execute correctly and create
the necessary picking. The order will be locked for any subsequent
operations as intended.

Tradeoff here;
We are sacrificing the universal application of one feature
(automatic lock on confirmed sales order) to fix a critical bug that
makes the entire FSM delivery process unusable. However, once the FSM
task is marked as done the SO linked to that task will be locked if the
global setting in Sales automatically locking the SO after confirmation
is enabled.

opw-4633977
opw-4749653
opw-4880664

closes odoo/enterprise#91905

X-original-commit: 5424c5a0c902b92e5c611b02a3c0a9b9a7760bd4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>